# backend/java/backend-java-spring/src/main/java/dev/multistack/app/dto/RegisterRequest.java
package dev.multistack.app.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank @Size(min = 3, max = 64) String username,
        @NotBlank @Size(min = 6, max = 128) String password
) {
}